Rename Reviewable infrastructure terminology
Having two different kinds of versions (major and minor) is not a good terminology. The term "minor version" implies that querying the non-latest minor version is a legit operation, but we actually don't want people to do that because the non-latest minor version is unapproved. Older minor versions are only interesting for history analysis.

We rename the terminology as follows:

 - "major version" -> "version"
 - "minor version" -> "adjustment